pyproject.toml
[build-system]
requires = [
    "setuptools==69.0.2",
    "wheel==0.42.0",
]
build-backend = "setuptools.build_meta"

[project]
version="1.24.1"
name = "icloudpd"
description = "icloudpd is a command-line tool to download photos and videos from iCloud."
readme = "README_PYPI.md"
requires-python = ">=3.8,<3.13"
keywords = ["icloud", "photo"]
license = {file="LICENSE.md"}
authors=[
    {name = "The iCloud Photo Downloader Authors"}
]
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "requests==2.31.0",
    "schema==0.7.5",
    "click==8.1.6",
    "tqdm==4.66.4",
    "piexif==1.1.3",
    "pyexiftool==0.5.6",
    "python-dateutil==2.9.0.post0",
    "urllib3==1.26.16",
    "typing_extensions==4.11.0",
    "Flask==3.0.3",
    "waitress==3.0.0",
    # from pyicloud_ipd
    "six==1.16.0",
    "tzlocal==5.1",
    "pytz==2024.1",
    "certifi==2022.12.7",
    "keyring==25.2.1",
    "keyrings-alt==5.0.1",
    "srp==1.0.21",
]

[project.optional-dependencies]
dev = [
    "twine==5.1.1",
    "pyinstaller==6.7.0",
    "wheel==0.42.0",
    "furo==2024.5.6",
    "Sphinx==7.3.7",
    "sphinx-autobuild==2024.4.16",
    "myst-parser==3.0.1"
]
devlinux = [
    "auditwheel==5.4.0",
    #"staticx==0.14.1",
    "scons==4.8.0"
]
test = [
    "pytest==7.4.3",
    "mock==5.1.0",
    "freezegun==1.5.1",
    "vcrpy==6.0.1",
    "pytest-cov==4.1.0",
    "ruff==0.5.0",
    "coveralls==3.3.1",
    "pytest-timeout==2.1.0",
    "pytest-xdist==3.3.1",
    "mypy==1.10.1",
    "types-python-dateutil==2.9.0.20241003",
    "types-pytz==2024.1.0.20240417",
    "types-tzlocal==5.1.0.1",
    "types-requests==2.31.0.2",
    "types-six==1.16.0",
    "types-urllib3==1.26.16",
    "types-tqdm==4.66.0.20240417",
    "types-mock==5.1.0.1",
    "types-waitress==3.0.0.20240423",
]

[project.urls]
Homepage="https://github.com/icloud-photos-downloader/icloud_photos_downloader"

[project.scripts]
icloudpd = "icloudpd.base:main"
icloud = "pyicloud_ipd.cmdline:main"

[tool.pytest.ini_options]
log_format = "%(levelname)-8s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
timeout = 300
testpaths = [
    "tests",
    "src" # needed for doctests    
]
pythonpath = [
    "src"
]
addopts = "--doctest-modules"

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)
exclude = ["starters"]

[[tool.mypy.overrides]]
module = ['piexif.*', 'vcr.*', 'srp.*', 'exiftool.*']
ignore_missing_imports = true

[tool.ruff]
line-length = 100
extend-exclude = [ "src/pyicloud_ipd" ]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # long lines
    "E501",
]

